Documenting oauth endpoints and allowing multo controllers
