+++ backend/tests/api/facebook.behavior.js
@@ -9,6 +9,7 @@ const sinon = require('sinon');
 const assert = require('assert');
 const nock = require('nock');
 const initObjectIdStub = require('objectid-stub');
+const passport = require('passport');
 const initRoutes = require('../../app/routes');
 
 describe('OAuth Facebook endpoints', () => {
@@ -32,6 +33,7 @@ describe('OAuth Facebook endpoints', () => {
       checkToken: sinon.stub().returns(true),
     };
     context.time = sinon.stub().returns(1664);
+    context.passport = passport;
     context.logger = {
       error: sinon.spy(),
       debug: sinon.spy(),
