+++ backend/tests/api/google.behavior.js
@@ -9,6 +9,7 @@ const sinon = require('sinon');
 const assert = require('assert');
 const nock = require('nock');
 const initObjectIdStub = require('objectid-stub');
+const passport = require('passport');
 
 const initRoutes = require('../../app/routes');
 
@@ -33,6 +34,7 @@ describe('OAuth Google endpoints', () => {
       checkToken: sinon.stub().returns(true),
     };
     context.time = sinon.stub().returns(1664);
+    context.passport = passport;
     context.logger = {
       error: sinon.spy(),
       debug: sinon.spy(),
