+++ backend/tests/workers/xee.behavior.js
@@ -6,6 +6,7 @@ const sinon = require('sinon');
 const assert = require('assert');
 const nock = require('nock');
 const initObjectIdStub = require('objectid-stub');
+const passport = require('passport');
 
 const xeeJobs = require('../../workers/xee/xee.jobs.js');
 
@@ -17,6 +18,7 @@ describe('Xee jobs', () => {
       env: { NODE_ENV: 'development' },
     };
     context.time = sinon.stub().returns(1664);
+    context.passport = passport;
     context.base = 'http://localhost/';
     context.logger = {
       error: sinon.spy(),
